iOS: Working on touch event handling
Handle finger down, motion, and up events to implement basic taps, long presses, and inertia scrolling.

Much finetuning still to be done, and certain widgets like input fields and scrollbars need a direct drag mode (they were working fine via the "mouse" events already).

+++ src/ui/sidebarwidget.c
@@ -322,7 +322,7 @@ static void updateItems_SidebarWidget_(iSidebarWidget *d) {
             iConstForEach(PtrArray, i, identities_GmCerts(certs_App())) {
                 const iGmIdentity *ident = i.ptr;
                 iSidebarItem *item = new_SidebarItem();
-                item->id = index_PtrArrayConstIterator(&i);
+                item->id = (uint32_t) index_PtrArrayConstIterator(&i);
                 item->icon = ident->icon;
                 set_String(&item->label, collect_String(subject_TlsCertificate(ident->cert)));
                 iDate until;
